Done I think but it probably has bugs still

+++ CardEditor/CardEditor/DeckMaker.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
 
 namespace CardEditor
 {
@@ -32,12 +33,17 @@ namespace CardEditor
         private static String name;
         private static String fileName;
 
-        private static int selected; // index of cardList selected, -1 means search is selected, -2 means name
+        private static int selected = -2; // index of cardList selected, -1 means search is selected, -2 means name
+        private static String searchTerm;
 
         private static readonly Rectangle listBox = new Rectangle((Game1.START_WIDTH - 500) / 2, (Game1.START_HEIGHT - 700) / 2, 500, 700);
+        private static readonly Rectangle searchBox = new Rectangle((Game1.START_WIDTH - 400) / 2, 120, 400, 30);
+        private static readonly Rectangle nameBox = new Rectangle((Game1.START_WIDTH - 300) / 2, 20, 300, 60);
 
         public static void New() {
             name = "Name";
+            fileName = null;
+            searchTerm = "";
 
             cardList = new List<CardWithCount>();
             LoadCards();
@@ -47,32 +53,82 @@ namespace CardEditor
         public static void Edit(String name) {
             SetupStats();
             LoadCards();
+            cardList = new List<CardWithCount>();
+            searchTerm = "";
 
             // load deck into cardList while updating stats
             DeckMaker.name = name;
+            fileName = name;
 
             String[] data = File.ReadAllLines("Content\\Decks\\" + name + ".deck");
 
             foreach(String line in data) {
                 String[] split = line.Split(',');
-
+                if(validCards.Contains(split[0])) {
+                    Add(new Card(split[0]), int.Parse(split[1]));
+                }
             }
-
-            // check for deleted cards
         }
 
         public static void Update() {
+            Text.Update();
+            Keys key = Text.GetCurrentKey();
+
             if(selected == -2) {
                 // change name
+                if(key == Keys.Back && name.Length > 0) {
+                    name = name.Substring(0, name.Length - 1);
+                } else {
+                    name += Text.KeyToLetter(key);
+                }
             }
             else if(selected == -1) {
-                // add new card
+                // type
+                // add new card from enter
+                if(key == Keys.Enter) {
+                    if(validCards.Contains(searchTerm)) {
+                        Add(new Card(searchTerm));
+                    }
+                }
+                else if(key == Keys.Tab) {
+                    searchTerm = "";
+                }
+                else if(key == Keys.Back && searchTerm.Length > 0) {
+                    searchTerm = searchTerm.Substring(0, searchTerm.Length - 1);
+                } else {
+                    searchTerm += Text.KeyToLetter(key);
+                }
+            }
+            else {
+                // an individual card is selected
+                if(key == Keys.OemPlus || key == Keys.Up) {
+                    Add(new Card(cardList[selected].Name));
+                }
+                else if(key == Keys.OemMinus || key == Keys.Down) {
+                    Remove(new Card(cardList[selected].Name));
+                }
             }
 
-
-            // select card for +/-
-
             // change selected item
+            if(Game1.JustClicked()) {
+                Point mousePos = Mouse.GetState().Position;
+                if(searchBox.Contains(mousePos)) {
+                    selected = -1;
+                }
+                else if(nameBox.Contains(mousePos)) {
+                    selected = -2;
+                }
+                else {
+                    // check list of cards
+                    List<Rectangle> cardBoxes = GetCardBoxes();
+                    for(int i = 0; i < cardBoxes.Count; i++) {
+                        if(cardBoxes[i].Contains(mousePos)) {
+                            selected = i;
+                            break;
+                        }
+                    }
+                }
+            }
         }
 
         public static void Draw(SpriteBatch sb) {
@@ -84,6 +140,20 @@ namespace CardEditor
             shrunk.Inflate(-4, -4);
             sb.Draw(Game1.Pixel, listBox, Color.White);
             sb.Draw(Game1.Pixel, shrunk, Color.Black);
+
+            List<Rectangle> listBoxes = GetCardBoxes();
+            if(selected == -2) {
+                sb.Draw(Game1.Pixel, nameBox, Color.LightBlue * 0.5f);
+            }
+            else if(selected >= 0) {
+                sb.Draw(Game1.Pixel, listBoxes[selected], Color.DarkBlue);
+            }
+
+            sb.Draw(Game1.Pixel, searchBox, selected == -1 ? Color.DarkBlue : Color.DarkGray);
+            DrawCentered(sb, searchTerm, new Vector2(searchBox.X + searchBox.Width / 2, searchBox.Y + searchBox.Height / 2));
+            for(int i = 0; i < cardList.Count; i++) {
+                DrawCentered(sb, cardList[i].Name + " x" + cardList[i].Count, new Vector2(listBoxes[i].X + listBoxes[i].Width / 2, listBoxes[i].Y + listBoxes[i].Height / 2));
+            }
 
             // draw stats
             Rectangle statsBox = new Rectangle(1150, (Game1.START_HEIGHT - 600) / 2, 400, 600);
@@ -116,6 +186,10 @@ namespace CardEditor
         }
 
         public static void Save() {
+            if(cardList.Count == 0) {
+                return;
+            }
+
             StreamWriter output = null; 
             try {
                 output = new StreamWriter("Content\\Decks\\" + name + ".deck");
@@ -124,8 +198,8 @@ namespace CardEditor
                     output.WriteLine(card.Name + "," + card.Count);
                 }
 
-                // delete old card if not overwritten
-                if (fileName != name) {
+                // delete old deck if not overwritten
+                if (fileName != null && fileName != name) {
                     File.Delete("Content\\Decks\\" + fileName + ".deck");
                     fileName = name;
                 }
@@ -140,12 +214,22 @@ namespace CardEditor
             }
         }
 
+        public static void Delete() {
+            File.Delete("Content\\Decks\\" + fileName + ".deck");
+        }
+
         private static void Add(Card card, int amount = 1) {
             // see if one exists already
             bool added = false;
             for(int i = 0; i < cardList.Count; i++) {
                 if(cardList[i].Name == card.Name) {
+                    // max of three copies
+                    if(cardList[i].Count >= 3) {
+                        return;
+                    }
+
                     cardList[i] = new CardWithCount(card.Name, cardList[i].Count + amount);
+                    added = true;
                     break;
                 }
             }
@@ -161,9 +245,36 @@ namespace CardEditor
             curve[card.Cost] += amount;
         }
 
+        private static void Remove(Card card) {
+            for(int i = 0; i < cardList.Count; i++) {
+                if(cardList[i].Name == card.Name) {
+                    cardList[i] = new CardWithCount(card.Name, cardList[i].Count - 1);
+
+                    if(cardList[i].Count <= 0) {
+                        // remove from list
+                        cardList.RemoveAt(i);
+                        selected = -1;
+                    }
+                    break;
+                }
+            }
+
+            // add to stats
+            total--;
+            stats[card.CardType]--;
+            curve[card.Cost]--;
+        }
+
         // get the names of all valid cards
         private static void LoadCards() {
+            String[] paths = Directory.GetFiles("Content\\Cards");
+            validCards = new List<String>();
 
+            foreach(String path in paths) {
+                String name = path.Substring(path.LastIndexOf('\\') + 1); // get rid of directories
+                name = name.Substring(0, name.IndexOf('.')); // get rid of .card
+                validCards.Add(name);
+            }
         }
 
         private static void SetupStats() {
@@ -182,6 +293,16 @@ namespace CardEditor
         private static void DrawCentered(SpriteBatch sb, String text, Vector2 center) {
             Vector2 dims = Game1.Font.MeasureString(text);
             sb.DrawString(Game1.Font, text, center - dims / 2, Color.White, 0f, Vector2.Zero, 1, SpriteEffects.None, 0);
+        }
+
+        private static List<Rectangle> GetCardBoxes() {
+            List<Rectangle> result = new List<Rectangle>();
+
+            for(int i = 0; i < cardList.Count; i++) {
+                result.Add(new Rectangle(searchBox.X, searchBox.Y + searchBox.Height * (i+1), searchBox.Width, searchBox.Height));
+            }
+
+            return result;
         }
     }
 }
